requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
PyPDF2
pydantic
